requirements.txt
amqp==2.6.1
anyio==3.6.2
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.2
billiard==3.6.4.0
black==23.1.0
bleach==6.0.0
blessed==1.19.1
blinker==1.4
bpython==0.22.1
Brlapi==0.7.0
celery==4.4.7
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.3
comm==0.1.2
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cryptography==2.8
cupshelpers==1.0
curtsies==0.4.1
cwcwidth==0.1.8
dbus-python==1.2.16
debugpy==1.6.6
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2
django-appconf==1.0.5
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-discover-runner==1.0
django-elasticsearch-dsl==7.3
django-environ==0.10.0
django-jazzmin==2.6.0
django-markdown==0.8.4
django-markdownfield==0.10.0
django-pytest==0.2.0
django-rest-framework==0.1.0
django-role-permissions==3.1.1
django-select2==8.1.1
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
drf-yasg2==1.19.4
duplicity==0.8.12.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
entrypoints==0.3
exceptiongroup==1.1.0
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.9.0
flake8==6.0.0
Flask==2.2.2
fqdn==1.5.1
future==0.18.2
greenlet==2.0.2
gunicorn==19.9.0
httplib2==0.14.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keyring==18.0.1
kombu==4.6.11
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
monotonic==1.5
mypy-extensions==1.0.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.10.4
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.7.0.72
packaging==23.0
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.6.1
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.16.2
pycodestyle==2.10.0
pycparser==2.21
pycups==1.9.73
pyflakes==3.0.1
Pygments==2.14.0
PyGObject==3.36.0
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyrsistent==0.19.3
pytest==7.2.2
python-apt==2.0.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-json-logger==2.0.4
pytz==2022.7.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.0.0
redis==4.1.4
reportlab==3.5.34
requests==2.28.2
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==2.3.1
Send2Trash==1.8.0
shortuuid==1.0.11
simplejson==3.16.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
systemd-python==234
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.4.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.14
usb-creator==0.3.7
uWSGI==2.0.21
vine==1.3.0
virtualenv==20.19.0
wadllib==1.3.3
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
whitenoise==6.4.0
wrapt==1.14.1
xkit==0.0.0
yapf==0.32.0
zipp==3.15.0
